Analytica/backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
greenlet
bcrypt
pydantic
pydantic-settings
alembic
python-multipart
python-jose[cryptography]
python-dotenv
psycopg2-binary
cryptography
metaapi-cloud-sdk
apscheduler
